+++ src/util/test/emptyChainDeployer.s.sol
@@ -6,8 +6,9 @@ import {MangroveDeployer} from "@mgv/script/core/deployers/MangroveDeployer.s.so
 import {OLKey} from "@mgv/src/core/MgvLib.sol";
 import {TestToken} from "@mgv/test/lib/tokens/TestToken.sol";
 import {MangroveOrderDeployer} from "@mgv-strats/script/strategies/mangroveOrder/deployers/MangroveOrderDeployer.s.sol";
+import {ActivateMangroveOrder} from "@mgv-strats/script/strategies/mangroveOrder/ActivateMangroveOrder.s.sol";
 import {KandelSeederDeployer} from "@mgv-strats/script/strategies/kandel/deployers/KandelSeederDeployer.s.sol";
-import {MangroveOrder} from "@mgv-strats/src/strategies/MangroveOrder.sol";
+import {MangroveOrder, RouterProxyFactory} from "@mgv-strats/src/strategies/MangroveOrder.sol";
 import {MgvReader} from "@mgv/src/periphery/MgvReader.sol";
 import {SimpleTestMaker} from "@mgv/test/lib/agents/TestMaker.sol";
 import {Mangrove} from "@mgv/src/core/Mangrove.sol";
@@ -16,6 +17,8 @@ import {Deployer} from "@mgv/script/lib/Deployer.sol";
 import {ActivateMarket, Market} from "@mgv/script/core/ActivateMarket.s.sol";
 import {PoolAddressProviderMock} from "@mgv-strats/script/toy/AaveMock.sol";
 import {IERC20} from "@mgv/lib/IERC20.sol";
+import {IPoolAddressesProvider} from "@mgv-strats/src/strategies/vendor/aave/v3/contracts/interfaces/IPoolAddressesProvider.sol";
+import {SimpleAaveLogic} from "@mgv-strats/src/strategies/routing_logic/SimpleAaveLogic.sol";
 
 /* 
 This script prepares a local chain for testing by mangrove.js.
@@ -29,6 +32,7 @@ contract EmptyChainDeployer is Deployer {
   address public weth;
   SimpleTestMaker public simpleTestMaker;
   MangroveOrder public mgo;
+  RouterProxyFactory public routerProxyFactory;
 
   function run() public {
     innerRun({gasprice: 1000, gasmax: 2_000_000, gasbot: broadcaster()});
@@ -38,7 +42,12 @@ contract EmptyChainDeployer is Deployer {
   function innerRun(uint gasprice, uint gasmax, address gasbot) public {
     MangroveDeployer mgvDeployer = new MangroveDeployer();
 
-    mgvDeployer.innerRun({chief: broadcaster(), gasprice: gasprice, gasmax: gasmax, gasbot: gasbot});
+    mgvDeployer.innerRun({
+      chief: broadcaster(),
+      gasprice: gasprice,
+      gasmax: gasmax,
+      gasbot: gasbot
+    });
 
     IMangrove mgv = mgvDeployer.mgv();
     MgvReader mgvReader = mgvDeployer.reader();
@@ -47,34 +56,67 @@ contract EmptyChainDeployer is Deployer {
     mgv.setUseOracle(false);
 
     broadcast();
-    tokenA = new TestToken({admin: broadcaster(), name: "Token A", symbol: "TokenA", _decimals: 18});
+    tokenA = new TestToken({
+      admin: broadcaster(),
+      name: "Token A",
+      symbol: "TokenA",
+      _decimals: 18
+    });
 
     broadcast();
     tokenA.setMintLimit(type(uint).max);
     fork.set("TokenA", address(tokenA));
 
     broadcast();
-    tokenB = new TestToken({admin: broadcaster(), name: "Token B", symbol: "TokenB", _decimals: 6});
+    tokenB = new TestToken({
+      admin: broadcaster(),
+      name: "Token B",
+      symbol: "TokenB",
+      _decimals: 6
+    });
 
     broadcast();
     tokenB.setMintLimit(type(uint).max);
     fork.set("TokenB", address(tokenB));
 
     broadcast();
-    dai = address(new TestToken({admin: broadcaster(), name: "DAI", symbol: "DAI", _decimals: 18}));
+    dai = address(
+      new TestToken({
+        admin: broadcaster(),
+        name: "DAI",
+        symbol: "DAI",
+        _decimals: 18
+      })
+    );
     fork.set("DAI", dai);
 
     broadcast();
-    usdc = address(new TestToken({admin: broadcaster(), name: "USD Coin", symbol: "USDC", _decimals: 6}));
+    usdc = address(
+      new TestToken({
+        admin: broadcaster(),
+        name: "USD Coin",
+        symbol: "USDC",
+        _decimals: 6
+      })
+    );
     fork.set("USDC", usdc);
 
     broadcast();
-    weth = address(new TestToken({admin: broadcaster(), name: "Wrapped Ether", symbol: "WETH", _decimals: 18}));
+    weth = address(
+      new TestToken({
+        admin: broadcaster(),
+        name: "Wrapped Ether",
+        symbol: "WETH",
+        _decimals: 18
+      })
+    );
     fork.set("WETH", weth);
 
     broadcast();
-    simpleTestMaker =
-      new SimpleTestMaker({_mgv: IMangrove(payable(mgv)), _ol: OLKey(address(tokenA), address(tokenB), 1)});
+    simpleTestMaker = new SimpleTestMaker({
+      _mgv: IMangrove(payable(mgv)),
+      _ol: OLKey(address(tokenA), address(tokenB), 1)
+    });
     fork.set("SimpleTestMaker", address(simpleTestMaker));
 
     ActivateMarket activateMarket = new ActivateMarket();
@@ -85,20 +127,44 @@ contract EmptyChainDeployer is Deployer {
     activateMarket.innerRun(mgv, mgvReader, Market(weth, dai, 1), 1e12, 1e12 / 1000, 0);
     activateMarket.innerRun(mgv, mgvReader, Market(weth, usdc, 1), 1e12, 1e12 / 1000, 0);
 
-    MangroveOrderDeployer mangroveDeployer = new MangroveOrderDeployer();
-    mangroveDeployer.innerRun({admin: broadcaster(), mgv: IMangrove(payable(mgv))});
-    MangroveOrder mangroveOrder = MangroveOrder(fork.get("MangroveOrder"));
     broadcast();
-    mangroveOrder.activate(dynamic([IERC20(tokenA), IERC20(tokenB)]));
+    routerProxyFactory = new RouterProxyFactory();
+    fork.set("RouterProxyFactory", address(routerProxyFactory));
 
-    address[] memory underlying = dynamic([address(tokenA), address(tokenB), dai, usdc, weth]);
+    MangroveOrderDeployer mgoeDeployer = new MangroveOrderDeployer();
+    mgoeDeployer.innerRun({
+      admin: broadcaster(),
+      mgv: IMangrove(payable(mgv)),
+      routerProxyFactory: routerProxyFactory
+    });
+    mgo = MangroveOrder(payable(fork.get("MangroveOrder")));
+    ActivateMangroveOrder activateMangroveOrder = new ActivateMangroveOrder();
+    activateMangroveOrder.innerRun({
+      mgvOrder: mgo,
+      tokens: dynamic([IERC20(tokenA), IERC20(tokenB)])
+    });
+
+    address[] memory underlying = dynamic(
+      [address(tokenA), address(tokenB), dai, usdc, weth]
+    );
     broadcast();
-    address aaveAddressProvider = address(new PoolAddressProviderMock(underlying));
+    address aaveAddressProvider = address(
+      new PoolAddressProviderMock(underlying)
+    );
+
+    // deploying logics
+    broadcast();
+    SimpleAaveLogic simpleAaveLogic = new SimpleAaveLogic(
+      IPoolAddressesProvider(aaveAddressProvider),
+      2 // variable interest rate mode
+    );
+    fork.set("SimpleAaveLogic", address(simpleAaveLogic));
+
 
     KandelSeederDeployer kandelSeederDeployer = new KandelSeederDeployer();
     kandelSeederDeployer.innerRun({
       mgv: IMangrove(payable(mgv)),
-      addressesProvider: aaveAddressProvider,
+      addressesProvider: IPoolAddressesProvider(aaveAddressProvider),
       aaveKandelGasreq: 629000, // see CoreKandelGasreqBaseTest
       kandelGasreq: 126000, // see CoreKandelGasreqBaseTest
       deployKandel: true,